pyproject.toml
[tool.poetry]
name = "vobla-vpn"
version = "0.1.0"
description = ""
authors = ["Eldos <eldos.baktybekov@gmail.com>"]
readme = "README.md"
packages = [{include = "vobla_vpn"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
httpx = "^0.23.0"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
